requirements.txt
fastapi
uvicorn
torch
spacy
transformers
python-docx
pymupdf
pydantic
sentencepiece
python-multipart